requirements.txt
connectonion>=0.3.0
playwright>=1.40.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-xdist>=3.0.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
Pillow
